remove cast to int

+++ game/properties/QuestionMark.java
@@ -16,7 +16,7 @@ public class QuestionMark extends Field{
             switch (cards[randNumber]){
                 case CASH -> player.setCash(player.getCash()+200);break;
                 case GOJAIL -> Prison.Imprisonment(player);break;//this method will be at Prison class
-                case TAX -> player.setCash((int) (0.9* player.getCash()));break;
+                case TAX -> player.setCash(0.9* player.getCash());break;
                 case MOVE -> player.setPosition(player.getPosition()+3);break;
                 case OUTJAIL -> player.setNoJail(player.getNoJail()+1);break;
                 case NOTAX -> player.setNoTax(player.getNoTax()+1);break;
